+++ source/georef/street_network.h
@@ -221,7 +221,7 @@ struct PathFinder {
 
 private:
     ///return the time the travel the distance at the current speed (used for projections)
-    navitia::time_duration crow_fly_duration(const double val) const;
+    navitia::time_duration crow_fly_duration(const float val) const;
 
     void add_custom_projections_to_path(Path& p, bool append_to_begin, const ProjectionData& projection, ProjectionData::Direction d) const;
 
@@ -276,7 +276,7 @@ struct StreetNetwork {
 Path create_path(const GeoRef& georef,
                  const std::vector<vertex_t>& reverse_path,
                  bool add_one_elt,
-                 double speed_factor);
+                 float speed_factor);
 
 /// Compute the angle between the last segment of the path and the next point
 int compute_directions(const navitia::georef::Path& path, const nt::GeographicalCoord& c_coord);
